Moved highlight into c.u.console.
The highlight function from c.m.utils is starting to get imported by
lots of modules outside of the c.maltego module. So to prevent ugly
circular imports it has been refactored to a more common place.
